--- src/components/PortfolioTracker/index.tsx
@@ -1,88 +1,0 @@
-const PortfolioTracker = () => {
-  return (
-    <div className="relative overflow-x-auto shadow-md rounded">
-      <div className="flex items-center justify-between flex-column flex-wrap md:flex-row space-y-4 md:space-y-0 pb-4 bg-white dark:bg-gray-900">
-        <div>
-          <button
-            className="inline-flex items-center text-gray-500 bg-white border border-gray-300 focus:outline-none hover:bg-gray-100 focus:ring-4 focus:ring-gray-200 font-medium rounded text-sm px-3 py-1.5 dark:bg-gray-800 dark:text-gray-400 dark:border-gray-600 dark:hover:bg-gray-700 dark:hover:border-gray-600 dark:focus:ring-gray-700"
-            type="button"
-          >
-            <span className="sr-only">Action button</span>
-            Add stock
-          </button>
-        </div>
-        <label htmlFor="table-search" className="sr-only">
-          Search
-        </label>
-        <div className="relative">
-          <div className="absolute inset-y-0 rtl:inset-r-0 start-0 flex items-center ps-3 pointer-events-none">
-            <svg
-              className="w-4 h-4 text-gray-500 dark:text-gray-400"
-              aria-hidden="true"
-              xmlns="http://www.w3.org/2000/svg"
-              fill="none"
-              viewBox="0 0 20 20"
-            >
-              <path
-                stroke="currentColor"
-                stroke-linecap="round"
-                stroke-linejoin="round"
-                stroke-width="2"
-                d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z"
-              />
-            </svg>
-          </div>
-          <input
-            type="text"
-            className="block p-2 ps-10 text-sm text-gray-900 border border-gray-300 rounded w-72 bg-gray-50 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
-            placeholder="Search stock"
-          />
-        </div>
-      </div>
-      <table className="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-200">
-        <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
-          <tr>
-            <th scope="col" className="px-6 py-3">
-              Symbol
-            </th>
-            <th scope="col" className="px-6 py-3">
-              Action
-            </th>
-            <th scope="col" className="px-6 py-3">
-              Buy Rate
-            </th>
-            <th scope="col" className="px-6 py-3">
-              Kitta
-            </th>
-            <th scope="col" className="px-6 py-3">
-              LTP
-            </th>
-            <th scope="col" className="px-6 py-3">
-              Total
-            </th>
-          </tr>
-        </thead>
-        <tbody>
-          <tr className="bg-white border-b dark:bg-gray-800 dark:border-gray-700 hover:bg-gray-50 dark:hover:bg-gray-600">
-            <td className="px-6 py-4">UPPER</td>
-            <td className="px-6 py-4">
-              <button className="font-medium text-blue-600 dark:text-blue-500 hover:underline">
-                Edit
-              </button>
-              <span className="px-2">|</span>
-              <button className="font-medium text-blue-600 dark:text-blue-500 hover:underline">
-                Delete
-              </button>
-            </td>
-            <td className="px-6 py-4">100</td>
-            <td className="px-6 py-4">100</td>
-            <td className="px-6 py-4">300</td>
-            <td className="px-6 py-4">30000</td>
-          </tr>
-        </tbody>
-      </table>
-    </div>
-  );
-};
-
-export default PortfolioTracker;
