+++ src/components/Table/index.ts
@@ -1,2 +1,3 @@
 export { default as AdvanceTable } from "./AdvanceTable";
+export { default as EditableTable } from "./EditableTable";
 export { default as SimpleTable } from "./SimpleTable";
