+++ src/components/index.ts
@@ -1,5 +1,5 @@
 export { default as About } from "./About";
-export { SimpleTable, AdvanceTable } from "./Table";
+export { SimpleTable, AdvanceTable, EditableTable } from "./Table";
 export { default as Breadcrumb } from "./Breadcrumb";
 export { default as Chart } from "./Chart";
 export { default as Contact } from "./Contact";
@@ -9,10 +9,11 @@ export { default as Login } from "./Login";
 export { default as MessageBox } from "./MessageBox";
 export { default as Navbar } from "./Navbar";
 export { default as NotFound } from "./NotFound";
-export { default as PortfolioTracker } from "./PortfolioTracker";
+export { default as PortfolioTracker } from "./Table/EditableTable";
 export { default as PrivacyPolicy } from "./PrivacyPolicy";
 export { default as Services } from "./Services";
 export { default as ShareCalculator } from "./ShareCalculator";
 export { default as Signup } from "./Signup";
 export { default as SimpleChart } from "./SimpleChart";
 export { default as StockDetail } from "./StockDetail";
+export { default as UserDataUpdate } from "./UserDataUpdate";
